Create High Level Wrapper for SoundEffects
